requirements.txt
annotated-types==0.5.0
anyio==3.7.1
azure-cognitiveservices-speech==1.30.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
decorator==4.4.2
exceptiongroup==1.1.2
fastapi==0.100.0
ffmpeg==1.4
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
moviepy==1.0.3
numpy==1.25.1
passlib==1.7.4
Pillow==10.0.0
proglog==0.1.10
pydantic==2.0.2
pydantic_core==2.1.2
pydub==0.25.1
PyJWT==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
webrtcvad==2.0.10
websockets==11.0.3
Werkzeug==2.3.6
zipp==3.16.0
